Fix `extractEntryTo` target
All `sanitize` calls take the `targetPath` and the `entryName` as arguments except for this one which actually prepend it with the `targetPath`. I couldn't find if there was a specific reason so far.

The result is that you might have the `targetPath` twice in the final `target`. Like `/Users/john/folder/Users/john/folder/file.txt` instead of the desired `/Users/john/folder/file.txt`.